requirements_simple.txt
# SUM v2 Requirements - Simple and Effective
# Total: 8 dependencies (down from 100+)

# Core Web Framework
flask==3.0.0
gunicorn==21.2.0

# Data Storage
redis==5.0.1
psycopg2-binary==2.9.9

# AI/ML (the heavy lifting)
transformers==4.36.0
torch==2.1.0

# Scientific Computing
numpy==1.24.3
scikit-learn==1.3.2